+++ backend/MySubs/MySubs.Domain/Models/Response/CheckMailUserResponse.cs
@@ -8,10 +8,10 @@ namespace MySubs.Domain.Models.Response
 {
     public class CheckMailUserResponse : ResponseResult
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
-        public static async Task<CheckMailUserResponse> Create(int id=0, string name="", string email="") => new CheckMailUserResponse
+        public static async Task<CheckMailUserResponse> Create(long id=0, string name="", string email="") => new CheckMailUserResponse
         {
             Id = id,
             Name = name,
